requirements.txt
kubernetes==10.0.1
azure_common==1.1.23
